+++ packages/synapse-bridge/package.json
@@ -17,7 +17,7 @@
 		"dev": "vite",
 		"build": "vue-tsc && vite build",
 		"preview": "vite preview",
-		"test": "jest"
+		"test": "jest --silent"
 	},
 	"dependencies": {
 		"vue": "3.3.4",
