+++ packages/synapse-bridge/src/mixins/customizable/types.d.ts
